+++ frontend/src/components/widgets/index.js
@@ -2,3 +2,8 @@ export { default as WidgetContainer } from "./WidgetContainer";
 export { default as WidgetHeader } from "./WidgetHeader";
 export { default as WidgetFooter } from "./WidgetFooter";
 export { default as WidgetMetric } from "./WidgetMetric";
+export { default as WidgetLoading } from "./WidgetLoading";
+export { default as WidgetErrorState } from "./WidgetErrorState";
+export { default as WidgetEmptyState } from "./WidgetEmptyState";
+export { default as WidgetMetricGrid } from "./WidgetMetricGrid";
+export { default as WidgetMetrics } from "./WidgetMetrics";