+++ frontend/src/features/portfolio/index.js
@@ -1,0 +1,6 @@
+export { default as usePortfolio } from "./hooks/usePortfolio";
+
+export { default as PortfolioSummaryMetrics } from "./components/PortfolioSummaryMetrics";
+export { default as PortfolioPerformanceMetrics } from "./components/PortfolioPerformanceMetrics";
+export { default as PortfolioStatisticsCard } from "./components/PortfolioStatisticsCard";
+export { default as PortfolioAllocationCard } from "./components/PortfolioAllocationCard";